move varint, bool, string, ustring and buffer to utils.js
