+++ src/datatypes/utils.js
@@ -1,0 +1,125 @@
+var assert = require('assert');
+
+var STRING_MAX_LENGTH = 240;
+var SRV_STRING_MAX_LENGTH = 32767;
+var getField= require("../utils").getField;
+
+module.exports = {
+    'varint': [readVarInt, writeVarInt, sizeOfVarInt],
+    'bool': [readBool, writeBool, 1],
+    'string': [readString, writeString, sizeOfString],
+    'ustring': [readString, writeString, sizeOfUString], // TODO : remove ustring
+    'buffer': [readBuffer, writeBuffer, sizeOfBuffer]
+};
+
+function readVarInt(buffer, offset) {
+    var result = 0;
+    var shift = 0;
+    var cursor = offset;
+
+    while (true) {
+        if (cursor + 1 > buffer.length) return null;
+        var b = buffer.readUInt8(cursor);
+        result |= ((b & 0x7f) << shift); // Add the bits to our number, except MSB
+        cursor++;
+        if (!(b & 0x80)) { // If the MSB is not set, we return the number
+            return {
+                value: result,
+                size: cursor - offset
+            };
+        }
+        shift += 7; // we only have 7 bits, MSB being the return-trigger
+        assert.ok(shift < 64, "varint is too big"); // Make sure our shift don't overflow.
+    }
+}
+
+function sizeOfVarInt(value) {
+    var cursor = 0;
+    while (value & ~0x7F) {
+        value >>>= 7;
+        cursor++;
+    }
+    return cursor + 1;
+}
+
+function writeVarInt(value, buffer, offset) {
+    var cursor = 0;
+    while (value & ~0x7F) {
+        buffer.writeUInt8((value & 0xFF) | 0x80, offset + cursor);
+        cursor++;
+        value >>>= 7;
+    }
+    buffer.writeUInt8(value, offset + cursor);
+    return offset + cursor + 1;
+}
+
+
+function readString (buffer, offset) {
+    var length = readVarInt(buffer, offset);
+    if (!!!length) return null;
+    var cursor = offset + length.size;
+    var stringLength = length.value;
+    var strEnd = cursor + stringLength;
+    if (strEnd > buffer.length) return null;
+
+    var value = buffer.toString('utf8', cursor, strEnd);
+    cursor = strEnd;
+
+    return {
+        value: value,
+        size: cursor - offset,
+    };
+}
+
+function writeString(value, buffer, offset) {
+    var length = Buffer.byteLength(value, 'utf8');
+    offset = writeVarInt(length, buffer, offset);
+    buffer.write(value, offset, length, 'utf8');
+    return offset + length;
+}
+
+
+function sizeOfString(value) {
+    var length = Buffer.byteLength(value, 'utf8');
+    assert.ok(length < STRING_MAX_LENGTH, "string greater than max length");
+    return sizeOfVarInt(length) + length;
+}
+
+function readBool(buffer, offset) {
+    if (offset + 1 > buffer.length) return null;
+    var value = buffer.readInt8(offset);
+    return {
+        value: !!value,
+        size: 1,
+    };
+}
+
+function writeBool(value, buffer, offset) {
+    buffer.writeInt8(+value, offset);
+    return offset + 1;
+}
+
+
+function readBuffer(buffer, offset, typeArgs, rootNode) {
+    var count = getField(typeArgs.count, rootNode);
+    return {
+        value: buffer.slice(offset, offset + count),
+        size: count
+    };
+}
+
+function writeBuffer(value, buffer, offset) {
+    value.copy(buffer, offset);
+    return offset + value.length;
+}
+
+function sizeOfBuffer(value) {
+    return value.length;
+}
+
+
+function sizeOfUString(value) {
+    var length = Buffer.byteLength(value, 'utf8');
+    assert.ok(length < SRV_STRING_MAX_LENGTH, "string greater than max length");
+    return sizeOfVarInt(length) + length;
+}
